Use backup downloaded files
We maintain a copy of libfabric, osu_benchmarks and ring_c
in /fsx, this can be used if there is a curl or git failure.
We need to make a local copy of this files, as they need to be
compiled and cannot be executed in /fsx as it might cause a race
condition. We only copy if we are unable to download these files

Signed-off-by: Dipti Kothari <dkothar@amazon.com>